New setting for audit

+++ ClearMessageOutlookAddIn/ClearMessageOutlookAddIn/SetupHelper.cs
@@ -4,8 +4,10 @@ using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Configuration.Install;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Xml;
@@ -28,8 +30,9 @@ namespace ClearMessageOutlookAddIn
 
                 //Saving the CustomActionData parameters to the install state dictionary to access afterwards in Install Commit
                 stateSaver.Add("TargetDir", Context.Parameters["targetdir"].ToString());
-                stateSaver.Add("BearerKey",Context.Parameters["bearerKey"].ToString());
+                stateSaver.Add("BearerKey", Context.Parameters["bearerKey"].ToString());
                 stateSaver.Add("ApiBaseUrl", Context.Parameters["apiBaseUrl"].ToString());
+                stateSaver.Add("AuditSetting", Context.Parameters["auditSetting"].ToString());
             }
             catch (Exception e)
             {
@@ -43,51 +46,74 @@ namespace ClearMessageOutlookAddIn
             {
                 base.Commit(savedState);
 
-                //Getting the location of the directory where the plugin files will get installed
-                string filePath = Path.GetDirectoryName(Context.Parameters["AssemblyPath"]);
+                bool isEmail = false;
 
-                //MessageBox.Show(filePath);
-
-                //Getting the settings.json file to update with bearer and endpoints
-                string jsonSettingsPath = filePath + "\\settings.json";
-
-                //MessageBox.Show(jsonSettingsPath);
-
-                SettingsModel settingsModel = new SettingsModel();
-
-                //Reading the settings.json file from the stream
-                string json = string.Empty;
-                using (StreamReader sr = new StreamReader(jsonSettingsPath))
+                if (!string.IsNullOrWhiteSpace(Context.Parameters["auditSetting"].ToString()))
                 {
-                    json = sr.ReadToEnd();
+                    
+                    string emailString = Context.Parameters["auditSetting"].ToString();
+                    isEmail = Regex.IsMatch(emailString, @"\A(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?)\Z", RegexOptions.IgnoreCase);
+
+                    if (isEmail)
+                    {
+                        MessageBox.Show("The audit setting contains the domain of email. The installation will stop and rollback.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        throw new Exception();
+                    }
+                    else
+                    {
+                        //Getting the location of the directory where the plugin files will get installed
+                        string filePath = Path.GetDirectoryName(Context.Parameters["AssemblyPath"]);
+
+                        //MessageBox.Show(filePath);
+
+                        //Getting the settings.json file to update with bearer and endpoints
+                        string jsonSettingsPath = filePath + "\\settings.json";
+
+                        //MessageBox.Show(jsonSettingsPath);
+
+                        SettingsModel settingsModel = new SettingsModel();
+
+                        //Reading the settings.json file from the stream
+                        string json = string.Empty;
+                        using (StreamReader sr = new StreamReader(jsonSettingsPath))
+                        {
+                            json = sr.ReadToEnd();
+                        }
+
+                        //MessageBox.Show(json);
+
+                        //Deserialzed the settings.json file to the SettingsModel object 
+                        settingsModel = JsonConvert.DeserializeObject<SettingsModel>(json);
+
+                        //MessageBox.Show(settings.ToString());
+                        //MessageBox.Show(Context.Parameters["bearerKey"]);
+                        //MessageBox.Show(Context.Parameters["apiBaseUrl"]);
+                        //MessageBox.Show(Context.Parameters["targetdir"]);
+
+                        //If settingsModel is not null then we will update the bearer token and endpoints
+                        if (settingsModel != null)
+                        {
+                            settingsModel.BearerKey = Context.Parameters["bearerKey"];
+                            settingsModel.ApiBaseUrl = Context.Parameters["apiBaseUrl"];
+                            settingsModel.AuditSetting = Context.Parameters["auditSetting"];
+                        }
+
+                        //MessageBox.Show(settingsModel.ToString());
+
+                        //Finally write and replace all the text in the settings.json file.
+                        File.WriteAllText(jsonSettingsPath, JsonConvert.SerializeObject(settingsModel));
+                        //MessageBox.Show("Done: " + jsonSettingsPath);
+                    }
                 }
-
-                //MessageBox.Show(json);
-
-                //Deserialzed the settings.json file to the SettingsModel object 
-                settingsModel = JsonConvert.DeserializeObject<SettingsModel>(json);
-
-                //MessageBox.Show(settings.ToString());
-                //MessageBox.Show(Context.Parameters["bearerKey"]);
-                //MessageBox.Show(Context.Parameters["apiBaseUrl"]);
-                //MessageBox.Show(Context.Parameters["targetdir"]);
-
-                //If settingsModel is not null then we will update the bearer token and endpoints
-                if (settingsModel != null)
+                else
                 {
-                    settingsModel.BearerKey = Context.Parameters["bearerKey"];
-                    settingsModel.ApiBaseUrl = Context.Parameters["apiBaseUrl"];
+                    MessageBox.Show("The audit setting cannot be empty. The installation will stop and rollback.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    throw new Exception();
                 }
-
-                //MessageBox.Show(settingsModel.ToString());
-
-                //Finally write and replace all the text in the settings.json file.
-                File.WriteAllText(jsonSettingsPath, JsonConvert.SerializeObject(settingsModel));
-                //MessageBox.Show("Done: " + jsonSettingsPath);
             }
             catch (Exception e)
             {
-                string s = e.Message;
+                base.Rollback(savedState);
             }
         }
     }
